Conflicts:
	src/components/transport_manager/src/bluetooth/bluetooth_transport_adapter.cc
	src/components/transport_manager/src/tcp/dnssd_service_browser.cc
	src/components/transport_manager/src/tcp/tcp_transport_adapter.cc
	src/components/transport_manager/src/transport_adapter/threaded_socket_connection.cc
	src/components/transport_manager/src/transport_adapter/transport_adapter_event.cc
	src/components/transport_manager/src/transport_adapter/transport_adapter_impl.cc
	src/components/transport_manager/src/transport_adapter/transport_adapter_listener_impl.cc
	src/components/transport_manager/src/transport_manager_impl.cc
	src/components/transport_manager/src/usb/libusb/usb_connection.cc

+++ src/components/utils/include/utils/file_system.h
@@ -55,7 +55,7 @@ uint64_t GetAvailableDiskSpace(const std::string& path);
  *
  * @param path to directory
  */
-uint32_t DirectorySize(const std::string& path);
+size_t DirectorySize(const std::string& path);
 
 /*
  * @brief Get size of current file
